+++ plugins/catalog-backend-module-github/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-catalog-backend-module-github
 
+## 0.2.2-next.0
+
+### Patch Changes
+
+- 70fa5ec3ec: Fixes the assignment of group member references in `GithubMultiOrgProcessor` so membership relations are resolved correctly.
+- 3280711113: Updated dependency `msw` to `^0.49.0`.
+- 754b5854df: Fix incorrectly exported GithubOrgEntityProvider as a type
+- Updated dependencies
+  - @backstage/catalog-client@1.2.0-next.0
+  - @backstage/plugin-catalog-backend@1.6.0-next.0
+  - @backstage/backend-common@0.16.1-next.0
+  - @backstage/integration@1.4.1-next.0
+  - @backstage/types@1.0.2-next.0
+  - @backstage/backend-plugin-api@0.1.5-next.0
+  - @backstage/plugin-catalog-node@1.2.2-next.0
+  - @backstage/backend-tasks@0.3.8-next.0
+  - @backstage/catalog-model@1.1.4-next.0
+  - @backstage/config@1.0.5-next.0
+  - @backstage/errors@1.1.4-next.0
+
 ## 0.2.0
 
 ### Minor Changes