+++ plugins/events-backend-module-aws-sqs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-events-backend-module-aws-sqs
 
+## 0.1.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-events-node@0.2.0-next.0
+  - @backstage/types@1.0.2-next.0
+  - @backstage/backend-plugin-api@0.1.5-next.0
+  - @backstage/backend-tasks@0.3.8-next.0
+  - @backstage/config@1.0.5-next.0
+
 ## 0.1.0
 
 ### Minor Changes